requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
fastapi==0.92.0
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.0.0
mangum==0.17.0
Markdown==3.4.1
pydantic==1.10.4
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
zipp==3.13.0
